[cleanup] remove adapter CLI and related code
The adapter CLI has been broken for awhile, and I think it will be a fair amount of work to fix. Focusing on having one CLI that works really well + does everything a Move dev would wish for seems more high-pri at the moment.

This deletes all the code related to the adapter CLI so we don't need to maintain it.