[objects] authority can return object layout to client in ObjectInfoRequest
More progress on interpreting object bytes inside the client.

- Modify the authority's ObjectInfoRequest to allow the client to (optionally) request the `MoveStructLayout` for an object in a variety of formats
- Modify the authority's ObjectInfoResponse to return the layout if requested
- Implement the appropriate traits in the authority to allow computing the layout from its local DB

Note: three approaches to this were discussed in https://github.com/GeunhwaJeong/fastnft/pull/397 and there was a general consensus that the third one (ask clients to store packages + use this to compute a layout locally) was preferred. However, after looking into this a bit, I concluded that this will be a major change to the client, which is not great because we need to start piping out JSON objects into the client service API's ASAP to unblock Geniteam (ideally today!).

In addition to the prioritization, I think approach (3) doesn't help a client who wants to get a structured view of an object that is not owned by them. Asking such a user to download every module related that object just to look at it + go through the dance of computing a layout from these modules doesn't seem great. Thus, I think exposing a way to ask an authority for the layout is reasonable even if we also implement approach (3).