fix: align methods for hash computation
Our current solution for hashing batches has 3 components:
1. it relies on never deserializing a batch, so that we hash the serialized message,
2. it is not resilient to hash confusion, lenght extension, etc (see #[87]).
3. it is split in two functions that disagree with each other (see #[188]).

This stopgap PR does nothing about 1 & 2, but realigns the `impl Hash for Batch` so that it matches with the rest of the code base, and uses it in processor tests.
THe security fixes of the batch hashing will be done in HaneulLabs/narwhal.

Fixes HaneulLabs/narwhal