feat: insert no-payload certificates as compressible
This adds an implementation of `Affiliated`, providing the insertion-time value of DAG vertexes' compressible field, which returns true on emtpy blocks.
Consequences:
- the DAG has genesis, but it's never reported in a read_causal,
- empty blocks get automatically purged by the next read_causal.

This can be seen as an alternative to HaneulLabs/narwhal, a continuation of HaneulLabs/narwhal.

This partially fixes HaneulLabs/narwhal in that the user will no longer have a reason to do `get_collections` or `remove_collections` on a genesis certificate.
(but that important detail needs to be better documented)