[haneul framework] add VecMap collection type
It's comment for Move programmers to want a collection with a map-like API, but the standard library does not have one.
This attempts to close that gap by providing a `VecMap`, which provides this API by using naive equality checking on keys.
As the comments on the collection suggest, this is not intended to be used for large collections--instead, the idea is to use this to replace some of the handrolling done in code like `validator_set::find_validator`. I expect there will be lots of third-party code that looks like this and want to provide a standard collection that pre-empts it.