genesis: completely build all objects at genesis contruction
Make genesis a bit more self contained by only having it contain
objects. This means that the publishing of the move and haneul frameworks
is done at the time genesis is constructed instead of requiring all
nodes to do this themselves. This makes genesis look like a "write-set
transaction" which may be a concept we add to Haneul at some point.