chore: update narwhal pointer
* chore: update narwhal pointer

Update the narwhal pointer and include the change where the
primary-to-primary interface was changed to use anemo (QUIC) vs gRPC.

This wasn't expected to be a breaking change but turned out to be one
due to the inability of an ipv4 udp socket from being able to send to an
ipv6 address. This is due to narwhal unconditionally binding on an ipv4
address (0.0.0.0) while the addresses included in the committee
structure are localhost which happens to first be resolved to ::1.

The simple short-term workaround is to change our config generation to
always use the ipv4 localhost address of 127.0.0.1.