[narwhal] fix cancellation safety for permit and future pattern
We have a pattern where we need to poll the FuturesUnordered and wait for receiving channel to have a permit at the same time.

It was implemented with `try_fut_and_permit` which introduced `join(permit(), waiting.next()`. This join is not cancellation safe (see  for more details) and can drop the message.

This PR fixes cancellation safety by chaining polling for the permit and polling `FuturesUnordered` which is cancellation safe.

To add a bit more flavor here - problem with previous implementation via join() is that join can resolve waiting.next() before permit is available and drops result, which is not safe. Polling permit first is safe because dropping permit 'reverts' the effects. (While dropping result from `FuturesUnordered` does not).

This PR also replaces macro with strongly typed `WithPermit` trait which seems like a more preferred approach - macro basically consumes anything that has method with name `reserve` which is not very sound.

Implementation side note - this PR introduce an `unwrap` when processing result from the waiting queue. This is same behaviour as `try_fut_and_permit` macro used to have, no strong opinion on how we actually *should* handle that.

Fixes