[BUGFIX][deps] Explicitly set patched version of sqlx
Haneul's dependency on `sqlx` needs to be overridden with a patched
version that uses a more recent version of `libsqlite3-sys`.  The
previous approach used a `patch` configuration in the root
`Cargo.toml`.

This works for builds within this project, but seems to break for
builds that originate from outside the workspace, but depend on one of
the workspace crates.

In this scenario, the patch doesn't apply on the dependency on `sqlx`
in `haneul-storage`, so it ends up conflicting with the dependency in
`workspace-hack`.

The most common example of this is someone depending on the Rust
SDK (`haneul-sdk`).

This diff ditches the `patch` and sets the dependency override
explicitly on `haneul-storage`.

Test Plan:

Haneul builds and runs:

```
haneul$ cargo build
haneul$ cargo simtest
haneul$ cargo nextest run
```

Dependents on the Haneul SDK build and run, e.g. following the
instructions at https://docs.haneul.io/build/rust-sdk succeeds.

Closes