update anemo to close connections after 30min ~ 90min
Background and details of the change is here: HaneulLabs/anemo. The goal is to mitigate memory growth in validators.

Anemo pointer is updated to current (87d60b) + the change above. The anemo pointer is not updated to latest main because there will be more risk, and there seems to be simtest failures.

The plan is to merge then pick this into the https://github.com/GeunhwaJeong/haneul/tree/releases/haneul-v0.15.0-release branch.