[ExecDriver] Launch execution driver with AuthorityState
After Gateway deprecation, since we plan to execute certificates with effects in execution driver, there should not be a case where AuthorityState is launched but execution driver is not needed. Keeping execution driver in activate authority no longer brings in benefits. But merging execution driver into AuthorityState has additional values:

1. More straight forward to wire TransactionManager to execution driver.
2. Less boilerplate related to metrics.

Since execution driver requires an Arc<> reference to AuthorityState, AuthorityState::new() is refactored to return Arc<AuthorityState> instead. This removes the need for callers to wrap AuthorityState wit Arc themselves.

This change is a prerequisite for , which uses TransactionManager for handle_certificate_with_effects() by default. Without running execution driver, a number of tests would fail.

An alternative approach is to launch ActiveAuthority when execution driver is needed, which is going to be almost all cases for tests and production code. The additional boilerplate seems unnecessary.