[reconfig] Wrap HaneulNode into Arc
Currently HaneulNode is created and passed around by value. To monitor
reconfiguration, it consumes the HaneulNode by value and wait forever. This
makes it impossible to write end-to-end tests that still need access to
the node with reconfiguration enabled.
This PR turns the reconfig channel into a Mutex, and then wrap the node
into an Arc for passing around.