Fix initialization of next_shared_object_versions
Previously, we attempted to initialize `next_shared_object_versions` the
first time we sequence a cert that touches a given object.
Unfortunately, this will not work if a transaction touching that object
has already been executed via checkpoint sync. See the referenced bug
below for details.

To fix this, we now initializate `next_shared_object_versions` for a
given object _before_ we set the shared locks for that object regardless
of whether we are setting the locks in the consensus task or from
checkpoint sync. This ensures that, if the object already existed (i.e.
was created in a prior epoch) we are guaranteed to read the latest
version of the object from the previous checkpoint.

However, because we are now initializing `next_shared_object_versions`
from multiple tasks, we need to account for races. Optimistic
transactions are used to handle races. When there is a race, one task
will succeed in initializing `next_shared_object_versions`, and then
other tasks will have to retry, will notice that
`next_shared_object_versions` has been initialized already, and will not
attempt to write to it again.

Also expose SetSnapshot function of rocksdb transactions.

Fixes