Distinguish insufficient gas errors
Before this PR, `InsufficientGas` is an enum in both
`ExecutionFailureStatus` and `HaneulError` but the meanings are different:
- `HaneulError::InsufficientGas` means the balance of gas object is lower
than the specified gas budget
- `ExecutionFailureStatus::InsufficientGas` means that the txn failed
b/c the actual gas cost has gone beyond `gas_budget`

As a result, sometimes it was a bit tricky to tell the exact error on
the client side if the type info of `ExecutionFailureStatus` or
`HaneulError` was missing. This PR renames the one in `HaneulError` to
`InsufficientGasForGasBudget` to distinguish them.