[Txn] Use a single write batch for transaction commit
We have observed in devnet that a node can restart in between updating
locks, and writing effects. Without effects, the transaction execution
gets retried. But with locks already updated, the transaction cannot
execute.

It seems the simplest way forward is to ensure updating objects / locks
and effects happen atomically. This is made simpler by the LockService
deprecation.

IIUC the issue may also be fixed with WAL, by retrying all of the
objects / locks / effects updates regardless of current db state. But it
seems a bit harder to be correct.