Minor changes to txn api
It might be a better idea to provide stronger isolation guarantee with
transaction api by default. This is the premise of this PR. It changes a
few small things:
1. By default the `rocksdb.transaction` will open the transaction with
set_snapshot set to true. If the user knows what they are doing, they
can use `rocksdb.transaction_without_snapshot` where they can choose to
call `get_for_update()` or set snapshot by themselves as part of the
transaction body. This makes the default mode run with stronger
isolation b/w concurrent transactions.
2. Renamed `TransactionWriteConflict` as `RetryableTransactionException`
which is returned on tx conflict i.e. `Busy` and when `TryAgain` is
returned from the db (as both are perfectly retryable situations).
`TryAgain` is returned when memtable history is not enough (because it
got flushed) to check for tx conflict and retrying a transaction is most
likely going to fix it.