Wallet validator unstaking1
I keep getting this error on devnet `Error: Error in getValidators:
Error: RPC Error: The response returned from RPC server does not match
the TypeScript definition. This is likely because the SDK version is not
compatible with the RPC server. Please update your SDK version to the
latest. ...` but works fine locally .





https://user-images.githubusercontent.com/8676844/212476961-e6bc2583-fd7e-4c95-9788-fad77c22416f.mov