[core] Refactor effects return value
This PR refactors the use of signed effects to prepare for finalized
transactions:
1. When we store SignedTransactionEffects, we should store
TrustedSignedTransactionEffects instead, and pass around
VerifiedSignedTransactionEffects since they are correct by construction.
This contributes to most of the refactoring in this PR. This includes
the executed_effects table as well as the wal.
2. `VerifiedHandleCertificateResponse` should contain
`VerifiedSignedTransactionEffects`, instead of
`SignedTransactionEffects`.
3. `VerifiedTransactionInfoResponse` should also contain
`VerifiedSignedTransactionEffects` instead of non-verified.
4. Remove the type template of `execute_certificate_with_effects`. It's
only called by fullnode today with an effects certs. Rename it to make
it clear, also add assert.
5. Remove `handle_transaction_and_effects_info_request` and
`execute_cert_to_true_effects` from the authority aggregator as they are
no longer used.