Disable collecting rocksdb perf metrics and fix db_name cardinality spike
A lot of RocksDB metrics are collected per db which started causing
cardinality spike because of ephemeral epoch databases (more pronounced
in private testnet where we are creating new epoch every 5 mins or so).
Specifically, the value for the label `db_name` now has multiple
possible values as epoch grows `epoch0`, `epoch1`, `epoch2`,.. To fix
this, we added a `MetricConf` struct which a user can pass while opening
the db and provide a label value to use for `db_name` label. After this
change, the value for label `db_name` will stay constant to `epoch`
(authorities epoch db) or `consensus_epoch` (consensus epoch db) even
with new epochs and stop the blowout.

The other problem is a lot of read/write/iter perf context metrics are
expensive because they are captured per column family. Given these are
not actively used, we are going to disable their collection for almost
all the databases (minus read perf context in authority store db). We
will turn them on if we are doing any perf related debugging.
NOTE: We are still going to collect db and cf metrics needed for db
operational insight (e.g disk and memory usage, etc) but with perf
metrics gone, these should be much fewer in number going forward.