[Narwhal] Always recover consensus dag on restart
When a node crashes, it is possible for consensus DAG to contain
certificates but have never committed any. In this case, the DAG still
needs to be recovered from the certificate store. Otherwise, there would
be a divergence between certificate store and DAG, which can result in
the DAG containing a certificate without necessary parents.

In another PR, I plan to harden consensus a bit on accepting incoming
certificates to ensure parents exist or have committed.

Closes