only add non-empty rewards as delegation when processing delegation switches
Right now when we process delegation switches at epoch changes, we add
the rewards portion of the stake as a delegation to the new validator
directly without checking whether the rewards are empty or not. When the
rewards are empty, it will cause the `request_add_delegation` code to
abort because `request_add_delegation` asserts that the new delegate
stake has balance > 0. This PR adds a check before we add the rewards as
new delegation, and a unit test for switching delegation when rewards ==
0.