remove another potential abort in staking_pool
This showed up when stress unit testing delegation code with random
amounts, random operations, etc. The rewards pool balance is 1 less than
the rewards we are trying to withdraw, possibly due to rounding. The
exact reason why this happened needs to be investigated. This PR removes
the possibility of aborting at `balance::split` when withdrawing rewards
from rewards pool.

UPDATE: switching the order in which we process new delegations and
delegation withdraws seems to fix the issue. It's still not super clear
whether the original order was the only trigger of the abort though.