[RWLocks] Correct casual order in checkpoints for read and write transactions.
We need to adjust how transactions are casually ordered in checkpoints
now that we have shared objects RWLocks.

With RWLocks we can have multiple transaction that depend on shared
object version N - many read transactions and single write transaction.
Those transactions depend on transaction that has written N, but they do
not depend on each other. And specifically, transaction that reads N and
writes N+1 does not depend on read-only transactions that also read N.
We do not add such read transactions to TransactionEffects of shared
object write transactions for next version to make sure
TransactionEffects do not grow too large (and because you do not need
read transactions to replay write transaction for next version).

However, when building checkpoints we supply transaction dependency tree
with additional dependency edges to make it look like write transaction
for next version casually depends on transactions that read previous
versions, for two reasons:

(1) Without this addition we could have peculiar checkpoints where
transaction reading version N appears after transaction that overwritten
this object with version N+1. This does not affect how transaction is
executed, but it is not something one would expect in casually ordered
list.

(2) On the practical side it will allow to simplify pruner as it can now
just tail checkpoints and delete objects in order they appear in
TransactionEffects::modified_at_versions in checkpoint.


On the philosophical level we have two slightly different types of
dependency graphs for transactions:

(a) `TransactionEffects::dependencies` describe a minimal dependency
tree of transactions, e.g. `TransactionEffects::dependencies` are
transaction digests that need to be executed before the transaction
itself.
(b) `Checkpoint` is a linearized list of executed transactions -
supposed the system was executing transactions sequentially, what would
be (one of) the orders of transaction that would yield the observed
result.

Those two DAGs were the same without RWLocks, but now that we have them
there is a slight difference in semantics.