[rpc] Add finalized effects
This PR adds the concept of finalized effects, to replace effects
certificate.
There are two different cases for a finalized effects, one is effects
certificate, and one is effects that were finalized in a specific
checkpoint.
This PR only touches the RPC interface to update it first. Latter we
will make validators to actually return it.