Use latest committee during bootstrap in embeeded reconfig observer
In the current setup when using embedded reconfig observer for remote
mode, we learn about the latest committee in the background but the
validator proxy is initialized with the genesis committee. This causes
initial gas requests to fail and eventually stress client to crash with
error. In this PR, we fix the behavior by learning about the latest
committee during validator proxy creation.