Shutdown server runtime before program exit
We don't ever terminate the server runtime in benchmark when we exit the
program. This causes RocksDB running in the server runtime to try to
access global state which is cleaned up during exit. This PR terminates
the server runtime with a oneshot channel and then exit.
Before this fix, I can reproduce the error in almost every run on my
linux computer. After this PR, I did not run into any errors even after
10+ runs.