[ts-sdk] Add `signMessage` and `signTransaction` APIs
The SDK does not currently support `signMessage` and `signTransaction`, so I added this missing APIs. I reworked how we construct intents and signed messages to clean up some duplication. I also changed the return types from these APIs to make them return the serialized signature string. The motivation for strings is that we're preparing something to be submitted over an API, and it ends up getting string-encoded at a later point anyway (often in the next function the data is passed to), so we might as well just make the return more sensible. This also lets these APIs be used over JSON-bridges without additional work (e.g. no need to do additional work to serialize / parse the data). Strings are also not mutable like Uint8Array which also should help denote that these things are finalized and ready for submission to the network.