crypto: Ed25519 fromSecretKey takes pure 32 bytes privkey
https://github.com/GeunhwaJeong/haneul/issues/6006

## What

This change makes `fromSecretKey` to take in an input of 32 bytes used
as the pure privkey before hashing and bit clamping. Previously it took
in a 64 byte privkey where the last 32 bytes is the public key and
optionally validate it. The `fromSeed` method is deprecated.

## Why
Taking in the 64 bytes privkey with baked in pubkey as the last 32 bytes
is an API subject to misuse. Especially with skipValidation = true.

## How to fix
If you run into errors probably because you are using 64 bytes privkey,
truncating it to the first 32 bytes (.slice(0, 32)) will just work.

## Test plan
Test added to confirm the haneul.keystore vs typescript are the same when
importing via mnenomics vs via raw secret key bytes. Also added an
example in comments since we got a lot of qs on how to import a haneul
keystore format to typescript