[TypeScript SDK] Fix TS definition for MoveCall
## Description 

We received a user reported bug that
https://explorer.haneul.io/transaction/BRL4LcY8hHytehMBPkNezXHQP2VNyG8pcWwzAGb16J9b
is not loading correctly.

This is because there's some mismatch between rust types and ts
definition:
- in rust, the arguments field can be optional if the vector is empty:
https://github.com/GeunhwaJeong/haneul/blame/main/crates/haneul-json-rpc-types/src/lib.rs#L1809-L1810
- in TS, the argument field is not optional:
https://github.com/GeunhwaJeong/haneul/blob/main/sdk/typescript/src/types/transactions.ts#L86

## Test Plan 

Verify that the transaction is loading correctly in preview:
https://explorer-git-fix-movecall-def-haneullabs-labs.vercel.app/transaction/BRL4LcY8hHytehMBPkNezXHQP2VNyG8pcWwzAGb16J9b

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [x] user-visible impact
- [x] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Make `arguments` field optional for MoveCall to match Rust definition.
This fixes a bug where the Explorer page does not load for transactions
with no argument.