Update default config values for timestamp and narwhal params
## Description 

Added additional logging for epoch duration/timestamp. And providing a
default value for `timestamp_ms` and updated defaults for nw params.
This will fix our deployments that were setting this value to 0 in
configs and triggering an instant epoch change from 0 -> 1. Will follow
up with an update to haneul-ops repo to remove setting of `timestamp_ms`
value in configs.

## Test Plan 

Confirmed epoch 0 duration in deployment in
[labnet](https://haneullabs.grafana.net/d/H9fvUo24z/haneul-epoch-metrics?orgId=1&var-Environment=haneullabs-metrics-internal&var-network=labnet&from=now-1h&to=now)