[6/x][programmable transactions] Calculate effects
- Adds the final bit of execution for determining the object changes and
collecting the user events

## Description 

This replicates the work currently done by the Move adapter, but will
apply to _all_ commands. Which in the future might call for a cleanup of
the TemporaryAuthorityStore

## Test Plan 

Not yet visible or usable

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes