[7/x][programmable transaction] Integrate into execution engine
## Description 

- Massaged some code to make it easier to integrate into the execution
engine
- Added the gas budget removal/refund logic
- We remove the max gas budget when we start executing, and put it back
at the end. This mimics setting that amount as "off limits" in the
course of the transaction. Without doing this, we might not have enough
to pay for gas at the end.
 
## Test Plan 

Nothing yet 

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes