[ts-sdk] Remove cached JSON RPC Provider
## Description 

This removes the cached RPC provider from the TypeScript SDK. This
hasn't been used in any of our application code, and I think there's
some fundamental challenges making it broadly used (e.g. what happens
when we serialize with an old object version? what is the recovery
mechanism?). Caching at this low level is hard to get right for all
use-cases, and we'd need to effectively build out a normalized caching
solution here to make this work well, which is better left to client
dapps to do (as-needed). Transaction serialization is also not the most
common case in dapps, which are significantly more read-heavy.

To solve the original use-case here (improving performance of
serializing transaction locally), we'll instead do this through a
batched API call in the new transaction builder, using some of the new
object APIs that don't return the full object data. We'll still allow
fully-resolved object references to be used in the event someone wants
to optimize things to avoid fetching.

## Test Plan 

Removed tests related to cached RPC provider.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [x] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

Removed cached JSON RPC Provider from the TypeScript SDK.