[upgrade] Add system state wrapper object
This PR introduces a wrapper object for the system state. The wrapper
object has ID 0x5, and all access to the system state object is done
indirectly through this wrapper object. Note that since this is already
a lot of refactoring, I am keeping the wrapper object simple in this PR:
it simply wraps the system state. In future PRs we will make the system
state object a dynamic field object of the wrapper.