[upgrade] Add EpochStaticInfo
## Description 

Currently we expose HaneulSystemState object directly both in the epoch
store cache as well as return to RPC calls (get_haneulSystemState). This
has two problems:
It confuses between the static epoch information that doesn't actually
change during the epoch (which is what we intended to cache), and the
actual dynamic system state. We are bound to make mistakes when these
two are mixed in use.
It couples the on-chain state data type and the RPC data type (as well
as implementation type inside epoch store). This will make it very
difficult to evolve the system state data structure down the road.
I am introducing a new type called EpochStaticInfo which captures all
the epoch static information from HaneulSystemState, and making it as flat
as possible (fields are at top-level as possible) so that it has almost
no structural dependencies. This will make it much much easier to evolve
and we no longer care about the specific changes to HaneulSystemState. We
can convert any version of HaneulSystemState into this type. The epoch
store will cache an instance of such static info, instead of the system
state directly. RPC interfaces will also return the epoch static info
(we can introduce a separate RPC API to return the latest system state
object from the database).

This PR does two things:
1. Introduce a new type called EpochStaticInfo, which captures all the
epoch static information from HaneulSystemState. It's currently not used
(otherwise this PR will grow out of control). Currently
ValidatorMetadata is used, but I also plan to flatten that.
2. Rename all API (gRPC, jsonRPC) names to reflect that they will be
returning epoch static info instead of system state.

## Test Plan 

build
---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ x] breaking change for a client SDKs
- [ x] breaking change for FNs (FN binary must upgrade)
- [ x] breaking change for validators or node operators (must upgrade
binaries)
- [ x] breaking change for on-chain data layout
- [ x] necessitate either a data wipe or data migration

### Release notes