Make VM, module cache (execution components) restart per epoch
Define an `ExectionComponent` struct which holds all instances related
to VM execution (and Move struct layout concepts).
Attach it to `AuthorityPerEpochStore` so it gets recreated on every
epoch change.
THe tricky part here has been to thread the `epoch_store` through a
bunch of functions to have it reach the places where the `module_cache`
was needed. Please take a look at that part because not all path could
be reached when a call to `load_epoch_store_one_call_per_task` was on
the stack. So thee are spurious and heavily commented calls to fetch the
`epoch_store`.

amnn haneullabsmark let's tlak offline how to test this properly, not too
sure what epoch changes tests we have