[haneul-framework] Remove non-delegation based staking flows
## Description 

This unifies the staking flows and staking representation for a validator so there is no
distinction between self and delegated stake. This additionally cleans
up a number of object fields that are no longer needed after this
change.

This is a breaking change since it changes the object representation for
`Validator` objects, as well as the validator staking/unstaking flows.

## Test Plan 

Updated existing framework tests
---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [X] user-visible impact
- [X] breaking change for a client SDKs
- [X] breaking change for FNs (FN binary must upgrade)
- [X] breaking change for validators or node operators (must upgrade
binaries)
- [X] breaking change for on-chain data layout
- [X] necessitate either a data wipe or data migration

### Release notes