[haneul-framework] Add test native to create OTWs
This adds a native function to `test_utils` that allows the creation of
one-time witnesses for testing purposes. This is then used to turn on a
test that couldn't be run previously since we didn't have an easy way of
creating OTWs.

## Test Plan 

Turns back on a test that needed to be turned off.