[wallet-ext] Rebuild wallet functionality on programmable transactions
## Description 

This reworks transaction construction in the wallet to use programmable
transactions. This was already happening under the hood, but this change
makes it happen explicitly now. As a nice side-effect, we were able to
get rid of double transactions on the stake flow.

I left out dapp transactions (which will come in a follow-up PR), and
the mint NFT (which is going away ).

## Test Plan 

Ran through core wallet flows.