move: recognize published-at field in Move.toml
## Description 

This registers a custom hook to recognize the `published-at` field in
package manifests. The value is unused currently, this is scaffolding
for collecting published addresses of dependencies (see
https://github.com/GeunhwaJeong/haneul/issues/8343).

Marking as non-user visible impact until this value is used (and where
release notes will accompany that PR).

## Test Plan 

Add test to recognize custom manifest properties.