Support for extensible committment schemes
## Description 

Allows multiple commitments to be added per checkpoint, and allows new
commitment types to be added in a forward compatible way.

Actually using multiple commitments will require consensus among
validators as to which commitment schemes should be used. That work is
TBD.

## Test Plan 

```cargo simtest```

---
If your changes are not user-facing and not a breaking change, you can skip the following section. Otherwise, please indicate what changed, and then add to the Release Notes section as highlighted during the release process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes