use different gas objects in dynamic committee test
## Description 

This test sometimes fails because we are sending multiple txns using the
same gas object and we may be hit with
`ObjectVersionUnavailableForConsumption ` error when the authority we
are querying hasn't executed the previous txn yet. Using different gas
objects solves the issue.

## Test Plan 

it's a test itself

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes