[tests] adding two tests to cover cases around Bullshark garbage collection
## Description 

The PR is introducing two new tests to cover cases around garbage
collection and especially when dealing with slow nodes as well. Also
refactored the leader election method when is used under testing, so we
can use it in a round robin fashion, instead of always having the first
only node doing the election.

## Test Plan 


---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes