[ts-sdk] Rewrite more SDK usage onto programmable tx
## Description 

This rewrites some of the usage of `SignableTransaction` to programmable
transactions in the SDK.

This also removes the `getGasObjectsOwnedByAddress` method, which
shouldn't be used anymore, and moves to a new internal helper for tests.

This PR also tweaks gas selection a little bit:
- Change gas selection to happen after input and command resolution. 
- Gas object selection now takes inputs into account, leading to less
transaction errors when using Haneul gas objects. This is mostly useful for
the existing transactions, generally we'd recommend using the gas object
+ split though.

## Test Plan 

Updated e2e tests.