Take db snapshots at epoch boundary and test snapshot restore
## Description 

This PR adds more databases that need chaeckpointing to enable quick
restore. We checkpoint committee store, checkpoint store and perpetual
while we take copy of epoch configuration in epoch db. Added db-tool
command for quick restore (but it is a trivial copy which could be
performed without it as well)

## Test Plan 

Tested restore from snapshot in a test by bringing up a full node from
it.