Object versioning system
## Description 

Prepares for future versioning of the object storage format.

Versioning process:

Object storage versioning is done lazily (at read time) - therefore we
must always preserve the
code for reading the very first storage version. For all versions, a
migration function

       f(V_n) -> V_(n+1)

must be defined. This way we can iteratively migrate the very oldest
version to the very newest
version at any point in the future.

To change the format of the object table value types (StoreObject and
StoreMoveObject), use the
following process:
- Add a new variant to the enum to store the new version type.
- Extend the `migrate` functions to migrate from the previous version to
the new version.
- Change `From<Object> for StoreObjectPair` to create the newest version
only.

Additionally, the first time we version these formats, we will need to:
- Add a check in the `TryFrom<StoreObjectPair> for Object` to see if the
object that was just
  read is the latest version.
- If it is not, use the migration function (as explained above) to
migrate it to the next
  version.
- Repeat until we have arrive at the current version.



## Test Plan 

`cargo simtest`

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [x] necessitate either a data wipe or data migration

### Release notes