[ts-sdk] Support programmable dev inspect
## Description 

This updates the TS SDK to support dev inspect programmable
transactions. Had to tweak some tests and skip one for now while a bug
is being fixed on the implementation side.

This was the final usecase for the legacy transaction serializer, so I
removed it. The TS SDK now only supports building programmable
transactions! 🎉

## Test Plan 

Updated tests.