[RPC][1 of 2] Add options to executeTransaction
## Description 

As a follow-up to https://github.com/GeunhwaJeong/haneul/pull/8888, this PR
changes the interface of `haneul_executeTransaction` without changing the
behavior(i.e., the `options` is ignored). To keep PRs small, I will
implement the logic to use `options` in the next PR, which will also
trigger more TS changes.

## Test Plan 

E2E tests

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [x] user-visible impact
- [x] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Update `executeTransaction` and `signAndExecuteTransaction` to take in
an additional parameter `HaneulTransactionResponse` which is used to
specify which fields to fetch (e.g., transaction, effects, events, etc).
By default, only the transaction digest will be returned.